BOSH and CF CLI are dead!
I haven't used the Cloud Foundry (CF) CLI or the BOSH (BOSH Outer SHell)
in over a year, and yet they've lingered in my Dockerfile.

This commit removes them from the Dockerfile, and renames the Docker
image to reflect that: fedora-golang-bosh -> fedora-golang.

It's bittersweet letting go of something that was once so central to my
life. But here we are.