Web page: consistently use first-person plural
In our second paragraph we switched between third-person plural and
first-person plural, and this really bothered me.

This commit fixes that by standardizing on first-person plural ("the
Royal We"). It sounds more familiar, less formal, and it blurs the line
between the service and the creators of the service, which I like.